# Optimizing Dijkstra: reducing dict lookups (up to 10% faster)